requirements.txt
certifi==2021.5.30
cryptography==3.4.8
Levenshtein==0.13.0
numpy==1.21.2
opencv-python==4.5.3.56
PyYAML==5.4.1
torch==1.9.0
typing-extensions==3.10.0.0